workers/image_optimizer/requirements.txt
Pillow>=11.0.0
numpy>=2.1.0
opencv-python-headless>=4.10.0.84
fastapi>=0.115.0
pydantic>=2.9.0
requests>=2.32.0
uvicorn[standard]>=0.32.0
